pyproject.toml
[tool.poetry]
name = "ms-alfred"
version = "0.1.0"
description = "Platzi Hackathon B2B Project"
authors = ["andresmaro <andresmaro@gmail.com>"]
readme = "README.md"
packages = [{include = "ms_alfred"}]

[tool.poetry.scripts]
start = "uvicorn main:app  --reload --host 0.0.0.0 --port 8000"

[tool.poetry.dependencies]
python = "^3.9"
jupyter = "^1.0.0"
fastapi = "^0.89.1"
uvicorn = "^0.20.0"
openai = "^0.26.1"
python-dotenv = "^0.21.0"
google-auth = "^2.16.0"
google-cloud-bigquery = "^3.4.2"
bigquery-python = "^1.15.0"
setuptools = "^66.0.0"
google-cloud-speech = "^2.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
